% Generated by roxygen2 (4.0.1): do not edit by hand
\name{easingLineChart}
\alias{easingLineChart}
\title{Credit easing data as an unstacked line chart.}
\usage{
easingLineChart(e)
}
\arguments{
\item{e}{an object of class \code{easing} as returned by \code{\link[easingr]{getEasingData}} and its many offspring.}
}
\description{
Provides a convenience function for passing an \code{easing} object to \code{xyplot}.
}
\details{
Provides several assumptions about the display of the \code{easing} data to correspond to similar presentations at the Cleveland Fed's data site.
}
\examples{
s <- getEasingSummary()
easingLineChart(s)
}
\seealso{
xyplot.easing easingAreaChart getEasingData
}

