% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{easingr}
\alias{easingr}
\alias{easingr-package}
\title{Cleveland Federal Reserve Bank credit easing policy tool data.}
\description{
Fetches the Federal Reserve Bank's credit easing policy tool data and transforms to time series.
}
\details{
Downloads the Cleveland Federal Reserve Bank's credit easing policy tool weekly data in its original CSV format, then translates that data into an \code{xts} time series object with easing product dollar value histories as columns.
}
\author{
Matt Barry \email{mrb@softisms.com}
}
\references{
Cleveland Fed research data site \url{http://www.clevelandfed.org/research/data/credit_easing/index.cfm}
}
\keyword{FRB}
\keyword{credit}
\keyword{easing}
\keyword{policy}

