% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getEasingCreditDepository}
\alias{getEasingCreditDepository}
\title{Get credit easing credit to depository institutions}
\usage{
getEasingCreditDepository(startDate = "012007", endDate = "122100")
}
\arguments{
\item{startDate}{query start date string in MMYYYY format, default 012007}

\item{endDate}{query end date string in MMYYYY format, default 122100}
}
\value{
A list of class \code{easing}
}
\description{
Downloads FRB credit easing policy credit to depository institutions data.
}
\details{
Downloads the Cleveland FRB data product for credit easing policy credit to depository institutions weekly time series, including columns for
\itemize{
\item primary credit
\item secondary credit
\item seasonal credit
}
}
\examples{
\dontrun{
cd <- getEasingCreditDepository()
head(cd$df)
}
}
\seealso{
getEasingData getEasingSummary getEasingDetails getEasingLending getEasingCreditExtensions getEasingProvidingLiquidity getEasingMaidenLane getEasingTraditionalHoldings  getEasingAgencyDebt getEasingLongTermPurchases
}

