% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getEasingCreditExtensions}
\alias{getEasingCreditExtensions}
\title{Get credit easing credit extensions data}
\usage{
getEasingCreditExtensions(startDate = "012007", endDate = "122100")
}
\arguments{
\item{startDate}{query start date string in MMYYYY format, default 012007}

\item{endDate}{query end date string in MMYYYY format, default 122100}
}
\value{
A list of class \code{easing}
}
\description{
Downloads FRB credit easing policy credit extensions data.
}
\details{
Downloads the Cleveland FRB data product for credit easing policy credit extensions weekly time series, including columns for
\itemize{
\item primary/other broker dealer
\item credit to AIG
\item other credit
}
}
\examples{
ce <- getEasingCreditExtensions()
head(ce$df)
}
\seealso{
getEasingData getEasingSummary getEasingDetails getEasingLending getEasingCreditDepository getEasingProvidingLiquidity getEasingMaidenLane getEasingTraditionalHoldings  getEasingAgencyDebt getEasingLongTermPurchases
}

