% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getEasingLending}
\alias{getEasingLending}
\title{Get credit easing lending.}
\usage{
getEasingLending(startDate = "012007", endDate = "122100")
}
\arguments{
\item{startDate}{query start date string in MMYYYY format, default 012007}

\item{endDate}{query end date string in MMYYYY format, default 122100}
}
\value{
A list of class \code{easing}
}
\description{
Downloads FRB credit easing policy lending data.
}
\details{
Downloads the Cleveland FRB data product for credit easing policy lending weekly time series, including columns for
\itemize{
\item repurchase agreements
\item credit to depository institutions
\item other fed assets
\item currency swaps
\item term auction credit
\item securities lent to dealers
\item credit extensions
}
}
\examples{
el <- getEasingLending()
head(el$df)
}
\seealso{
getEasingData getEasingSummary getEasingDetails getEasingCreditDepository getEasingCreditExtensions getEasingProvidingLiquidity getEasingMaidenLane getEasingTraditionalHoldings  getEasingAgencyDebt getEasingLongTermPurchases
}

