% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getEasingLongTermPurchases}
\alias{getEasingLongTermPurchases}
\title{Get credit easing long-term treasury purchases data.}
\usage{
getEasingLongTermPurchases(startDate = "012007", endDate = "122100")
}
\arguments{
\item{startDate}{query start date string in MMYYYY format, default 012007}

\item{endDate}{query end date string in MMYYYY format, default 122100}
}
\value{
A list of class \code{easing}
}
\description{
Downloads FRB credit easing policy tools long-term treasury purchases data.
}
\details{
Downloads the Cleveland FRB data product for credit easing policy tools long-term treasury purchases weekly time series, including columns for
\itemize{
\item treasury purchases
}
}
\examples{
lt <- getEasingLongTermPurchases()
head(lt$df)
}
\seealso{
getEasingData getEasingSummary getEasingDetails getEasingLending getEasingCreditDepository getEasingCredtExtensions getEasingProvidingLiquidity getEasingMaidenLane  getEasingTraditionalHoldings getEasingAgencyDebt
}

