% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getEasingMaidenLane}
\alias{getEasingMaidenLane}
\title{Get credit easing credit Maiden Lane data.}
\usage{
getEasingMaidenLane(startDate = "012007", endDate = "122100")
}
\arguments{
\item{startDate}{query start date string in MMYYYY format, default 012007}

\item{endDate}{query end date string in MMYYYY format, default 122100}
}
\value{
A list of class \code{easing}
}
\description{
Downloads FRB credit easing policy tools Maiden Lane data.
}
\details{
Downloads the Cleveland FRB data product for credit easing policy tools Maiden Lane weekly time series, including columns for
\itemize{
\item Maiden Lane 1
\item Maiden Lane 2
\item Maiden Lane 3
}
}
\examples{
\dontrun{
ml <- getEasingMaidenLane()
head(ml$df)
}
}
\seealso{
getEasingData getEasingSummary getEasingDetails getEasingLending getEasingCreditDepository getEasingCredtExtensions getEasingProvidingLiquidity getEasingTraditionalHoldings  getEasingAgencyDebt getEasingLongTermPurchases
}

