% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getEasingSummary}
\alias{getEasingSummary}
\title{Get credit easing summary.}
\usage{
getEasingSummary(startDate = "012007", endDate = "122100")
}
\arguments{
\item{startDate}{query start date string in MMYYYY format, default 012007}

\item{endDate}{query end date string in MMYYYY format, default 122100}
}
\value{
A list of class \code{easing}
}
\description{
Downloads FRB credit easing policy summary data.
}
\details{
Downloads the Cleveland FRB data product for credit easing policy summary weekly time series, including columns for
\itemize{
\item traditional security holdings
\item long-term treasury purchases
\item lending to financial institutinos
\item liquidity to key credit markets
\item federal agency debt mortgage backed securities purchases
}
}
\examples{
es <- getEasingSummary()
head(es$df)
}
\seealso{
getEasingData getEasingDetails getEasingLending getEasingCreditDepository getEasingCreditExtensions getEasingProvidingLiquidity getEasingMaidenLane getEasingTraditionalHoldings  getEasingAgencyDebt getEasingLongTermPurchases
}

