% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getEasingTraditionalHoldings}
\alias{getEasingTraditionalHoldings}
\title{Get credit easing credit traditional security holdings data.}
\usage{
getEasingTraditionalHoldings(startDate = "012007", endDate = "122100")
}
\arguments{
\item{startDate}{query start date string in MMYYYY format, default 012007}

\item{endDate}{query end date string in MMYYYY format, default 122100}
}
\value{
A list of class \code{easing}
}
\description{
Downloads FRB credit easing policy tools traditional security holdings data.
}
\details{
Downloads the Cleveland FRB data product for credit easing policy tools traditional security holdings weekly time series, including columns for
\itemize{
\item traditional security holdings
}
}
\examples{
\dontrun{
th <- getEasingTraditionalHoldings()
head(th$df)
}
}
\seealso{
getEasingData getEasingSummary getEasingDetails getEasingLending getEasingCreditDepository getEasingCredtExtensions getEasingProvidingLiquidity getEasingMaidenLane  getEasingAgencyDebt getEasingLongTermPurchases
}

